requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

PyYAML
pandas
pyarrow
torch
torchvision
timm
scikit-learn
Pillow
tqdm
numpy
transformers
accelerate
qwen-vl-utils
gradio

